test: consolidate receipt safety coverage
